@@ -1,10 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AssemblyName>Naos.Application.Web</AssemblyName>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <AssemblyName>Naos.Application.Web3</AssemblyName>
     <RootNamespace>Naos.Application.Web</RootNamespace>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
   </PropertyGroup>
 
   <ItemGroup>
@@ -20,12 +19,12 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AspNetCore.HealthChecks.UI" Version="2.2.34" />
-    <PackageReference Include="AspNetCore.HealthChecks.Uris" Version="2.2.3" />
-    <PackageReference Include="NSwag.AspNetCore" Version="13.0.6" />
-    <PackageReference Include="NSwag.Generation" Version="13.0.6" />
-    <PackageReference Include="NJsonSchema" Version="10.0.23" />
-    <PackageReference Include="Serilog.AspNetCore" Version="2.1.1" />
+    <PackageReference Include="AspNetCore.HealthChecks.UI" Version="3.0.0" />
+    <PackageReference Include="AspNetCore.HealthChecks.Uris" Version="3.0.0" />
+    <PackageReference Include="NSwag.AspNetCore" Version="13.1.2" />
+    <PackageReference Include="NSwag.Generation" Version="13.1.2" />
+    <PackageReference Include="NJsonSchema" Version="10.0.24" />
+    <PackageReference Include="Serilog.AspNetCore" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
